+++ src/test/resources/journeys.conf
@@ -24,7 +24,7 @@ journeys {
     # The parts your journey is made of. A part is made one or more requests.
   aa-journey = {
       description = "AA / LTA calculation user journey"
-      load = 0.5
+      load = 1
       parts = [
         Basic-flow,
         Annual-Allowance-flow,
@@ -39,6 +39,8 @@ journeys {
         Annual-Allowance-2023,
         Lifetime-Allowance,
         Calculate-results,
+        Auth-wizard,
+        Submission-route
       ]
     }
 }
