+++ src/test/resources/services-local.conf
@@ -40,4 +40,19 @@ services {
       host = localhost
       port = 12805
     }
+    auth-login-stub {
+      protocol = http
+      host = localhost
+      port = 9949
+    }
+    auth-session-stub {
+      protocol = http
+      host = localhost
+      port = 12802
+    }
+    final-submit-stub {
+      protocol = http
+      host = localhost
+      port = 12803
+    }
 }
