+++ src/test/resources/services.conf
@@ -36,8 +36,25 @@ services {
     port = 443
   }
   submit-stub{
-      protocol = https
-      host = www.staging.tax.service.gov.uk
-      port = 443
-    }
+    protocol = https
+    host = www.staging.tax.service.gov.uk
+    port = 443
+  }
+
+  auth-login-stub {
+    protocol = https
+    host = www.staging.tax.service.gov.uk
+    port = 443
+  }
+
+  auth-session-stub {
+    protocol = https
+    host = calculate-public-pension-adjustment.protected.mdtp
+    port = 443
+  }
+  final-submit-stub {
+    protocol = http
+    host = www.staging.tax.service.gov.uk
+    port = 443
+  }
 }
