+++ src/test/scala/uk/gov/hmrc/perftests/calculatefrontend/requests/calculateAA/AA2016PostRequests.scala
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.perftests.calculatefrontend.requests
+package uk.gov.hmrc.perftests.calculatefrontend.requests.calculateAA
 
 import io.gatling.core.Predef._
 import io.gatling.http.Predef._
 import io.gatling.http.request.builder.HttpRequestBuilder
 import uk.gov.hmrc.perftests.calculatefrontend.Configuration
 
-object AA2022Requests extends Configuration {
+object AA2016PostRequests extends Configuration {
 
   val calculateRoute: String                            = s"$calculationUrl/public-pension-adjustment"
-  val period2022: String                                = "/2022"
+  val period2016Post: String                            = "/2016-post"
   val scheme1: String                                   = "/0"
   val scheme2: String                                   = "/1"
-  val scheme3: String                                   = "/2"
   val whatYouWillNeedAaPageUrl: String                  = "/what-you-will-need-aa"
   val memberMoreThanOnePensionPageUrl: String           = "/member-more-than-one-pension"
   val pensionSchemeDetailsPageUrl: String               = "/pension-scheme-details"
@@ -42,67 +41,54 @@ object AA2022Requests extends Configuration {
   val whoPaidAnnualAllowanceChargePageUrl: String       = "/who-paid-annual-allowance-charge"
   val howMuchYouPayChargePageUrl: String                = "/how-much-you-pay-charge"
   val totalIncomePageUrl: String                        = "/total-income"
-  val thresholdIncomePageUrl: String                    = "/threshold-income"
 
   val navigateToWhatYouWillNeedAaPage: HttpRequestBuilder =
-    http("Navigate to whatYouWillNeedAaPageUrl page " + period2022)
-      .get(calculateRoute + whatYouWillNeedAaPageUrl + period2022)
+    http("Navigate to whatYouWillNeedAaPageUrl page " + period2016Post)
+      .get(calculateRoute + whatYouWillNeedAaPageUrl + period2016Post)
       .check(status.is(200))
 
   val navigateToMemberMoreThanOnePensionPage: HttpRequestBuilder =
-    http("Navigate to memberMoreThanOnePensionPageUrl page " + period2022)
-      .get(calculateRoute + memberMoreThanOnePensionPageUrl + period2022)
+    http("Navigate to memberMoreThanOnePensionPageUrl page " + period2016Post)
+      .get(calculateRoute + memberMoreThanOnePensionPageUrl + period2016Post)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitMemberMoreThanOnePensionConfirmation(value: String): HttpRequestBuilder =
     http("memberMoreThanOnePensionPageUrl : " + value)
-      .post(calculateRoute + memberMoreThanOnePensionPageUrl + period2022)
+      .post(calculateRoute + memberMoreThanOnePensionPageUrl + period2016Post)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToWhichSchemeDetailsPage: HttpRequestBuilder =
-    http("Navigate to whichSchemeDetailsPage page " + period2022)
-      .get(calculateRoute + whichSchemeDetailsPageUrl + period2022 + scheme1)
+    http("Navigate to whichSchemeDetailsPage page " + period2016Post)
+      .get(calculateRoute + whichSchemeDetailsPageUrl + period2016Post + scheme1)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitWhichSecondSchemeDetailsConfirmation(value: String): HttpRequestBuilder =
     http("whichSecondSchemeDetailsPage : " + value)
-      .post(calculateRoute + whichSchemeDetailsPageUrl + period2022 + scheme2)
+      .post(calculateRoute + whichSchemeDetailsPageUrl + period2016Post + scheme2)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToWhichSecondSchemeDetailsPage: HttpRequestBuilder =
-    http("Navigate to whichSchemeDetailsPage page " + period2022)
-      .get(calculateRoute + whichSchemeDetailsPageUrl + period2022 + scheme2)
+    http("Navigate to whichSchemeDetailsPage page " + period2016Post)
+      .get(calculateRoute + whichSchemeDetailsPageUrl + period2016Post + scheme2)
       .check(status.is(200))
       .check(saveCsrfToken)
 
-  val navigateToWhichThirdSchemeDetailsPage: HttpRequestBuilder =
-    http("Navigate to whichSchemeDetailsPage page " + scheme3)
-      .get(calculateRoute + whichSchemeDetailsPageUrl + period2022 + scheme3)
-      .check(status.is(200))
-      .check(saveCsrfToken)
-
-  def submitWhichThirdSchemeDetailsConfirmation(value: String): HttpRequestBuilder =
-    http("whichSecondSchemeDetailsPage : " + value)
-      .post(calculateRoute + whichSchemeDetailsPageUrl + period2022 + scheme3)
-      .formParam("csrfToken", "${csrfToken}")
-      .formParam("value", value)
-      .check(status.is(303))
-  def submitWhichSchemeDetailsConfirmation(value: String): HttpRequestBuilder      =
+  def submitWhichSchemeDetailsConfirmation(value: String): HttpRequestBuilder =
     http("whichSchemeDetailsPage : " + value)
-      .post(calculateRoute + whichSchemeDetailsPageUrl + period2022 + scheme1)
+      .post(calculateRoute + whichSchemeDetailsPageUrl + period2016Post + scheme1)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToFirstPensionSchemeInputAmountsPageUrlPage: HttpRequestBuilder =
-    http("Navigate to pensionSchemeInputAmounts page " + period2022 + " Scheme " + scheme1)
-      .get(calculateRoute + pensionSchemeInputAmountsPageUrl + period2022 + scheme1)
+    http("Navigate to pensionSchemeInputAmounts page " + period2016Post + " Scheme " + scheme1)
+      .get(calculateRoute + pensionSchemeInputAmountsPageUrl + period2016Post + scheme1)
       .check(status.is(200))
       .check(saveCsrfToken)
 
@@ -111,69 +97,55 @@ object AA2022Requests extends Configuration {
     revisedPIA: String
   ): HttpRequestBuilder =
     http("pensionSchemeInputAmounts : " + originalPIA + " " + revisedPIA)
-      .post(calculateRoute + pensionSchemeInputAmountsPageUrl + period2022 + scheme1)
+      .post(calculateRoute + pensionSchemeInputAmountsPageUrl + period2016Post + scheme1)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("originalPIA", originalPIA)
       .formParam("revisedPIA", revisedPIA)
       .check(status.is(303))
 
   val navigateToDidYouPayAChargePage: HttpRequestBuilder =
-    http("Navigate to DidYouPayACharge page " + period2022 + " Scheme " + scheme1)
-      .get(calculateRoute + didYouPayAChargePageUrl + period2022 + scheme1)
+    http("Navigate to DidYouPayACharge page " + period2016Post + " Scheme " + scheme1)
+      .get(calculateRoute + didYouPayAChargePageUrl + period2016Post + scheme1)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitDidYouPayAChargeConfirmation(value: String): HttpRequestBuilder =
     http("DidYouPayACharge : " + value)
-      .post(calculateRoute + didYouPayAChargePageUrl + period2022 + scheme1)
+      .post(calculateRoute + didYouPayAChargePageUrl + period2016Post + scheme1)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToAddAnotherSchemePage: HttpRequestBuilder =
-    http("Navigate to AddAnotherScheme page " + period2022 + " Scheme " + scheme1)
-      .get(calculateRoute + addAnotherSchemePageUrl + period2022 + scheme1)
+    http("Navigate to AddAnotherScheme page " + period2016Post + " Scheme " + scheme1)
+      .get(calculateRoute + addAnotherSchemePageUrl + period2016Post + scheme1)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitAddAnotherSchemeConfirmation(value: String): HttpRequestBuilder =
     http("AddAnotherScheme : " + value)
-      .post(calculateRoute + addAnotherSchemePageUrl + period2022 + scheme1)
+      .post(calculateRoute + addAnotherSchemePageUrl + period2016Post + scheme1)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToSecondPensionSchemeDetailsPage: HttpRequestBuilder =
-    http("Navigate to pensionSchemeDetails page " + period2022 + " Scheme " + scheme2)
-      .get(calculateRoute + pensionSchemeDetailsPageUrl + period2022 + scheme2)
+    http("Navigate to pensionSchemeDetails page " + period2016Post + " Scheme " + scheme2)
+      .get(calculateRoute + pensionSchemeDetailsPageUrl + period2016Post + scheme2)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitPensionSecondSchemeDetailsConfirmation(schemeName: String, schemeTaxRef: String): HttpRequestBuilder =
     http("pensionSchemeDetails : " + schemeName + " " + schemeTaxRef)
-      .post(calculateRoute + pensionSchemeDetailsPageUrl + period2022 + scheme2)
-      .formParam("csrfToken", "${csrfToken}")
-      .formParam("schemeName", schemeName)
-      .formParam("schemeTaxRef", schemeTaxRef)
-      .check(status.is(303))
-
-  val navigateToThirdPensionSchemeDetailsPage: HttpRequestBuilder =
-    http("Navigate to pensionSchemeDetails page " + period2022 + " Scheme " + scheme3)
-      .get(calculateRoute + pensionSchemeDetailsPageUrl + period2022 + scheme3)
-      .check(status.is(200))
-      .check(saveCsrfToken)
-
-  def submitPensionThirdSchemeDetailsConfirmation(schemeName: String, schemeTaxRef: String): HttpRequestBuilder =
-    http("pensionSchemeDetails : " + schemeName + " " + schemeTaxRef)
-      .post(calculateRoute + pensionSchemeDetailsPageUrl + period2022 + scheme3)
+      .post(calculateRoute + pensionSchemeDetailsPageUrl + period2016Post + scheme2)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("schemeName", schemeName)
       .formParam("schemeTaxRef", schemeTaxRef)
       .check(status.is(303))
 
   val navigateToSecondPensionSchemeInputAmountsPageUrlPage: HttpRequestBuilder =
-    http("Navigate to pensionSchemeInputAmounts page " + period2022 + " Scheme " + scheme2)
-      .get(calculateRoute + pensionSchemeInputAmountsPageUrl + period2022 + scheme2)
+    http("Navigate to pensionSchemeInputAmounts page " + period2016Post + " Scheme " + scheme2)
+      .get(calculateRoute + pensionSchemeInputAmountsPageUrl + period2016Post + scheme2)
       .check(status.is(200))
       .check(saveCsrfToken)
 
@@ -182,174 +154,118 @@ object AA2022Requests extends Configuration {
     revisedPIA: String
   ): HttpRequestBuilder =
     http("pensionSchemeInputAmounts : " + originalPIA + " " + revisedPIA)
-      .post(calculateRoute + pensionSchemeInputAmountsPageUrl + period2022 + scheme2)
-      .formParam("csrfToken", "${csrfToken}")
-      .formParam("originalPIA", originalPIA)
-      .formParam("revisedPIA", revisedPIA)
-      .check(status.is(303))
-
-  val navigateToThirdPensionSchemeInputAmountsPageUrlPage: HttpRequestBuilder =
-    http("Navigate to pensionSchemeInputAmounts page " + period2022 + " Scheme " + scheme2)
-      .get(calculateRoute + pensionSchemeInputAmountsPageUrl + period2022 + scheme2)
-      .check(status.is(200))
-      .check(saveCsrfToken)
-
-  def submitPensionThirdSchemeInputAmountsPageUrlConfirmation(
-    originalPIA: String,
-    revisedPIA: String
-  ): HttpRequestBuilder =
-    http("pensionSchemeInputAmounts : " + originalPIA + " " + revisedPIA)
-      .post(calculateRoute + pensionSchemeInputAmountsPageUrl + period2022 + scheme2)
+      .post(calculateRoute + pensionSchemeInputAmountsPageUrl + period2016Post + scheme2)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("originalPIA", originalPIA)
       .formParam("revisedPIA", revisedPIA)
       .check(status.is(303))
 
   val navigateToDidYouPayAChargeSecondSchemePage: HttpRequestBuilder =
-    http("Navigate to DidYouPayACharge page " + period2022 + " Scheme " + scheme2)
-      .get(calculateRoute + didYouPayAChargePageUrl + period2022 + scheme2)
+    http("Navigate to DidYouPayACharge page " + period2016Post + " Scheme " + scheme2)
+      .get(calculateRoute + didYouPayAChargePageUrl + period2016Post + scheme2)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitDidYouPayAChargeSecondSchemeConfirmation(value: String): HttpRequestBuilder =
     http("DidYouPayACharge : " + value)
-      .post(calculateRoute + didYouPayAChargePageUrl + period2022 + scheme2)
-      .formParam("csrfToken", "${csrfToken}")
-      .formParam("value", value)
-      .check(status.is(303))
-
-  val navigateToDidYouPayAChargeThirdSchemePage: HttpRequestBuilder =
-    http("Navigate to DidYouPayACharge page " + period2022 + " Scheme " + scheme3)
-      .get(calculateRoute + didYouPayAChargePageUrl + period2022 + scheme3)
-      .check(status.is(200))
-      .check(saveCsrfToken)
-
-  def submitDidYouPayAChargeThirdSchemeConfirmation(value: String): HttpRequestBuilder =
-    http("DidYouPayACharge : " + value)
-      .post(calculateRoute + didYouPayAChargePageUrl + period2022 + scheme3)
+      .post(calculateRoute + didYouPayAChargePageUrl + period2016Post + scheme2)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToWhoPaidAnnualAllowanceChargePage: HttpRequestBuilder =
-    http("Navigate to whoPaidAnnualAllowanceCharge page " + period2022 + " Scheme " + scheme1)
-      .get(calculateRoute + whoPaidAnnualAllowanceChargePageUrl + period2022 + scheme1)
+    http("Navigate to whoPaidAnnualAllowanceCharge page " + period2016Post + " Scheme " + scheme1)
+      .get(calculateRoute + whoPaidAnnualAllowanceChargePageUrl + period2016Post + scheme1)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitWhoPaidAnnualAllowanceChargeConfirmation(value: String): HttpRequestBuilder =
     http("whoPaidAnnualAllowanceCharge : " + value)
-      .post(calculateRoute + whoPaidAnnualAllowanceChargePageUrl + period2022 + scheme1)
+      .post(calculateRoute + whoPaidAnnualAllowanceChargePageUrl + period2016Post + scheme1)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToHowMuchYouPayChargePage: HttpRequestBuilder =
-    http("Navigate to howMuchYouPayCharge page " + period2022 + " Scheme " + scheme1)
-      .get(calculateRoute + howMuchYouPayChargePageUrl + period2022 + scheme1)
+    http("Navigate to howMuchYouPayCharge page " + period2016Post + " Scheme " + scheme1)
+      .get(calculateRoute + howMuchYouPayChargePageUrl + period2016Post + scheme1)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitHowMuchYouPayChargeConfirmation(value: String): HttpRequestBuilder =
     http("howMuchYouPayCharge : " + value)
-      .post(calculateRoute + howMuchYouPayChargePageUrl + period2022 + scheme1)
+      .post(calculateRoute + howMuchYouPayChargePageUrl + period2016Post + scheme1)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToAddAnotherSchemeTwoPage: HttpRequestBuilder =
-    http("Navigate to AddAnotherScheme page " + period2022 + " Scheme " + scheme2)
-      .get(calculateRoute + addAnotherSchemePageUrl + period2022 + scheme2)
+    http("Navigate to AddAnotherScheme page " + period2016Post + " Scheme " + scheme2)
+      .get(calculateRoute + addAnotherSchemePageUrl + period2016Post + scheme2)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitAddAnotherSchemeTwoConfirmation(value: String): HttpRequestBuilder =
     http("AddAnotherScheme : " + value)
-      .post(calculateRoute + addAnotherSchemePageUrl + period2022 + scheme2)
-      .formParam("csrfToken", "${csrfToken}")
-      .formParam("value", value)
-      .check(status.is(303))
-
-  val navigateToAddAnotherSchemeThreePage: HttpRequestBuilder =
-    http("Navigate to AddAnotherScheme page " + period2022 + " Scheme " + scheme3)
-      .get(calculateRoute + addAnotherSchemePageUrl + period2022 + scheme3)
-      .check(status.is(200))
-      .check(saveCsrfToken)
-
-  def submitAddAnotherSchemeThreeConfirmation(value: String): HttpRequestBuilder =
-    http("AddAnotherScheme : " + value)
-      .post(calculateRoute + addAnotherSchemePageUrl + period2022 + scheme3)
+      .post(calculateRoute + addAnotherSchemePageUrl + period2016Post + scheme2)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToContributedOtherDbDcSchemePage: HttpRequestBuilder =
-    http("Navigate to contributedOtherDbDcScheme page " + period2022)
-      .get(calculateRoute + contributedOtherDbDcSchemePageUrl + period2022)
+    http("Navigate to contributedOtherDbDcScheme page " + period2016Post)
+      .get(calculateRoute + contributedOtherDbDcSchemePageUrl + period2016Post)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitContributedOtherDbDcSchemeConfirmation(value: String): HttpRequestBuilder =
     http("contributedOtherDbDcScheme : " + value)
-      .post(calculateRoute + contributedOtherDbDcSchemePageUrl + period2022)
+      .post(calculateRoute + contributedOtherDbDcSchemePageUrl + period2016Post)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToWhichContributedDuringRemedyPeriodPage: HttpRequestBuilder =
-    http("Navigate to whichContributedDuringRemedyPeriodPageUrl page " + period2022)
-      .get(calculateRoute + whichContributedDuringRemedyPeriodPageUrl + period2022)
+    http("Navigate to whichContributedDuringRemedyPeriodPageUrl page " + period2016Post)
+      .get(calculateRoute + whichContributedDuringRemedyPeriodPageUrl + period2016Post)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitWhichContributedDuringRemedyPeriodConfirmation(value: String): HttpRequestBuilder =
     http("whichContributedDuringRemedyPeriodPageUrl : " + value)
-      .post(calculateRoute + whichContributedDuringRemedyPeriodPageUrl + period2022)
+      .post(calculateRoute + whichContributedDuringRemedyPeriodPageUrl + period2016Post)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value[1]", value)
       .check(status.is(303))
 
   val navigateToPiaForDbPensionPage: HttpRequestBuilder =
-    http("Navigate to piaForDbPension page " + period2022)
-      .get(calculateRoute + piaForDbPensionPageUrl + period2022)
+    http("Navigate to piaForDbPension page " + period2016Post)
+      .get(calculateRoute + piaForDbPensionPageUrl + period2016Post)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitPiaForDbPensionConfirmation(value: String): HttpRequestBuilder =
     http("piaForDbPensionPageUrl : " + value)
-      .post(calculateRoute + piaForDbPensionPageUrl + period2022)
+      .post(calculateRoute + piaForDbPensionPageUrl + period2016Post)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToTotalIncomePageUrlPage: HttpRequestBuilder =
-    http("Navigate to totalIncome page " + period2022)
-      .get(calculateRoute + piaForDbPensionPageUrl + period2022)
+    http("Navigate to totalIncome page " + period2016Post)
+      .get(calculateRoute + piaForDbPensionPageUrl + period2016Post)
       .check(status.is(200))
       .check(saveCsrfToken)
 
   def submitTotalIncomePageUrlConfirmation(value: String): HttpRequestBuilder =
     http("totalIncomePage : " + value)
-      .post(calculateRoute + piaForDbPensionPageUrl + period2022)
+      .post(calculateRoute + piaForDbPensionPageUrl + period2016Post)
       .formParam("csrfToken", "${csrfToken}")
       .formParam("value", value)
       .check(status.is(303))
 
   val navigateToCheckYourAnswersPeriodPage: HttpRequestBuilder =
-    http("Navigate to checkYourAnswersPeriod page " + period2022)
-      .get(calculateRoute + checkYourAnswersPeriodPageUrl + period2022)
+    http("Navigate to checkYourAnswersPeriod page " + period2016Post)
+      .get(calculateRoute + checkYourAnswersPeriodPageUrl + period2016Post)
       .check(status.is(200))
-
-  val navigateToThresholdIncomePage: HttpRequestBuilder =
-    http("Navigate to thresholdIncome page " + period2022)
-      .get(calculateRoute + thresholdIncomePageUrl + period2022)
-      .check(status.is(200))
-
-  def submitThresholdIncomePageConfirmation(value: String): HttpRequestBuilder =
-    http("thresholdIncome : " + value)
-      .post(calculateRoute + thresholdIncomePageUrl + period2022)
-      .formParam("csrfToken", "${csrfToken}")
-      .formParam("value", value)
-      .check(status.is(303))
-
 }
